Create README.md
The doubly linked list is a very efficient data structure for any programming language, so do in C++; it is used for the manipulation of elements. The doubly linked list is made up of many nodes in a sequence, and it further provides pointer concepts for performing various operations on top of those nodes present within the list.